+++ packages/react/src/components/Panel/index.tsx
@@ -6,6 +6,9 @@ import { useStore } from '../../hooks/useStore';
 import type { ReactFlowState } from '../../types';
 
 export type PanelProps = HTMLAttributes<HTMLDivElement> & {
+  /** Set position of the panel
+   * @example 'top-left' | 'top-center' | 'top-right' | 'bottom-left' | 'bottom-center' | 'bottom-right'
+   */
   position?: PanelPosition;
   children: ReactNode;
 };