+++ packages/svelte/CHANGELOG.md
@@ -8,6 +8,8 @@
 - add slot to `Controls`
 - cleanup `ControlButton` types
 - infer types for `getIncomers`, `getOutgoers`, `updateEdge`, `addEdge` and `getConnectedEdges` thanks @joeyballentine
+- refactor handles: prefix with flow id for handling nested flows
+- add comments for types like `SvelteFlowProps` or `Node` for a better developer experience
 
 ## 0.0.33